Reorganizacja testów jednostkowych
