+++ Solution1/CarHire.UnitTests/CarHire.UnitTests.csproj
@@ -95,14 +95,18 @@
     </When>
     <Otherwise>
       <ItemGroup>
-        <Reference Include="Microsoft.VisualStudio.QualityTools.UnitTestFramework" />
+        <Reference Include="Microsoft.VisualStudio.QualityTools.UnitTestFramework">
+          <Private>False</Private>
+        </Reference>
       </ItemGroup>
     </Otherwise>
   </Choose>
   <ItemGroup>
     <Compile Include="App_Start\NinjectWebCommon.cs" />
-    <Compile Include="UnitTest1.cs" />
+    <Compile Include="Tests\PagintionTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Tests\CarTest.cs" />
+    <Compile Include="Tests\UserTest.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
